# api-gateway/cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing users, tweets, comments, likes, and direct messages.",
        "title": "API Gateway",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5050",
    "basePath": "/",
    "paths": {
...
        "/likes/comment/user/{user_id}": {
            "get": {
                "description": "Get all likes for a specific comment by a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "likes"
                ],
                "summary": "Get likes for a comment by user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api-gateway_protos_like-proto.GetLikeCommentByUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_like-handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_like-handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/likes/comment/{comment_id}": {
            "get": {
                "description": "Get all likes for a specific comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "likes"
                ],
                "summary": "Get likes for a comment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comment ID",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api-gateway_protos_like-proto.GetLikesCommentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_like-handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_like-handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/likes/comment/{comment_id}/{user_id}": {
            "post": {
                "description": "Like a specific comment by user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "likes"
                ],
                "summary": "Like a comment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Comment ID",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api-gateway_protos_like-proto.CreateLikeCommentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_like-handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_like-handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove like from a specific comment by user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "likes"
                ],
                "summary": "Unlike a comment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Comment ID",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api-gateway_protos_like-proto.DeleteLikeCommentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_like-handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_like-handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/likes/tweet/{tweet_id}/{user_id_to_get}": {
            "get": {
                "description": "Get all likes for a specific tweet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "likes"
                ],
                "summary": "Get likes for a tweet",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tweet ID",
                        "name": "tweet_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "User ID to get likes",
                        "name": "user_id_to_get",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api-gateway_protos_like-proto.GetLikesTweetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_like-handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_like-handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/likes/tweet/{tweet_id}/{user_id}": {
            "post": {
                "description": "Like a specific tweet by user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "likes"
                ],
                "summary": "Like a tweet",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Tweet ID",
                        "name": "tweet_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api-gateway_protos_like-proto.CreateLikeTweetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_like-handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_like-handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove like from a specific tweet by user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "likes"
                ],
                "summary": "Unlike a tweet",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Tweet ID",
                        "name": "tweet_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api-gateway_protos_like-proto.DeleteLikeTweetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_like-handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_like-handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/likes/tweet/{tweet_id}/{user_id}/{like_identifier}": {
            "get": {
                "description": "Get a specific liked tweet by user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "likes"
                ],
                "summary": "Get a liked tweet",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Tweet ID",
                        "name": "tweet_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Like Identifier",
                        "name": "like_identifier",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api-gateway_protos_like-proto.GetLikedTweetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_like-handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_like-handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/likes/user/{user_id}": {
            "get": {
                "description": "Get all tweets liked by a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "likes"
                ],
                "summary": "Get liked tweets by user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api-gateway_protos_like-proto.GetLikeTweetByUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_like-handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_like-handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api-gateway_protos_like-proto.CreateLikeCommentResponse": {
            "type": "object",
            "properties": {
                "like": {
                    "$ref": "#/definitions/api-gateway_protos_like-proto.Like"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api-gateway_protos_like-proto.CreateLikeTweetResponse": {
            "type": "object",
            "properties": {
                "like": {
                    "$ref": "#/definitions/api-gateway_protos_like-proto.Like"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api-gateway_protos_like-proto.DeleteLikeCommentResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api-gateway_protos_like-proto.DeleteLikeTweetResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api-gateway_protos_like-proto.GetLikeCommentByUserResponse": {
            "type": "object",
            "properties": {
                "likes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api-gateway_protos_like-proto.Like"
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api-gateway_protos_like-proto.GetLikeTweetByUserResponse": {
            "type": "object",
            "properties": {
                "likes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api-gateway_protos_like-proto.Like"
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api-gateway_protos_like-proto.GetLikedTweetResponse": {
            "type": "object",
            "properties": {
                "likes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api-gateway_protos_like-proto.Like"
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api-gateway_protos_like-proto.GetLikesCommentResponse": {
            "type": "object",
            "properties": {
                "likes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api-gateway_protos_like-proto.Like"
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api-gateway_protos_like-proto.GetLikesTweetResponse": {
            "type": "object",
            "properties": {
                "likes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api-gateway_protos_like-proto.Like"
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api-gateway_protos_like-proto.Like": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "like_identifier": {
                    "type": "string"
                },
                "liked_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "internal_handlers_like-handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    }
}